+++ app/javascript/vue/tasks/collecting_events/filter/components/Filter.vue
@@ -65,7 +65,7 @@ const WITH_PARAMS = [
   'collection_objects',
   'collectors',
   'data_attributes',
-  'images',
+  'depictions',
   'geographic_area',
   'georeferences',
   'identifiers',